+++ pkg/deploy/generator/util.go
@@ -1,0 +1,11 @@
+package generator
+
+// Copyright (c) Microsoft Corporation.
+// Licensed under the Apache License 2.0.
+
+func max(a, b int) int {
+	if a > b {
+		return a
+	}
+	return b
+}
